+++ lib/private/data-stages/entities.lua
@@ -49,30 +49,6 @@ end
 
 --@ category_name
 --@ entity_name
---@ to_add_mask_name
-function krastorio.entities.addCollisionMaskOnEntity(category_name, entity_name, to_add_mask_name)
-  local entity = krastorio.entities.getEntity(category_name, entity_name)
-  if entity and to_add_mask_name ~= nil then
-    local collision_mask = collision_mask_util.get_mask(entity)
-      or { "item-layer", "object-layer", "player-layer", "water-tile" }
-    local miss = true
-    for _, mask_name in pairs(collision_mask) do
-      if mask_name == to_add_mask_name then
-        miss = false
-        break
-      end
-    end
-    if miss then
-      table.insert(collision_mask, to_add_mask_name)
-      entity.collision_mask = collision_mask
-    end
-    return true
-  end
-  return false
-end
-
---@ category_name
---@ entity_name
 --@ to_add_crafting_category
 function krastorio.entities.addCraftingCategory(category_name, entity_name, to_add_crafting_category)
   local entity = krastorio.entities.getEntity(category_name, entity_name)