+++ prototypes/resources/mineral-water.lua
@@ -68,8 +68,4 @@ data:extend({
     richness = true,
     category = "resource",
   },
-  {
-    type = "noise-layer",
-    name = "mineral-water",
-  },
 })