+++ prototypes/vanilla-changes/optional/damage-and-ammo.lua
@@ -269,7 +269,7 @@ if krastorio.general.getSafeSettingValue("kr-damage-and-ammo") then
       tile_name = "nuclear-ground",
       radius = 12,
       apply_projection = true,
-      tile_collision_mask = { layers = { ["water-tile"] = true } },
+      tile_collision_mask = { layers = { water_tile = true } },
     },
     {
       type = "destroy-cliffs",
@@ -463,7 +463,7 @@ if krastorio.general.getSafeSettingValue("kr-damage-and-ammo") then
             {
               type = "create-entity",
               entity_name = "nuclear-smouldering-smoke-source",
-              tile_collision_mask = { layers = { ["water-tile"] = true } },
+              tile_collision_mask = { layers = { water_tile = true } },
             },
           },
         },