+++ prototypes/vanilla-changes/optional/radar-changes.lua
@@ -199,7 +199,7 @@ if krastorio.general.getSafeSettingValue("kr-rebalance-radar") then
       icon = kr_technologies_icons_path .. "advanced-radar.png",
       icon_size = 256,
       icon_mipmaps = 4,
-      prerequisites = { "kr-radar", "advanced-electronics" },
+      prerequisites = { "kr-radar", "advanced-circuit" },
       effects = {
         {
           type = "unlock-recipe",